Refactor dependency naming
